v 1.6 - Added PList entry to stop getting nagged about export compliance by apple.  Fixed crash when deleting an account.  Allow for zero balance.  Lots of bug fixes.

+++ ErgoIOSLiteClient/SwaggerModels/Transactions.swift
@@ -16,8 +16,8 @@ public struct ErgoTransaction: Codable {
     public var dataInputs: [ErgoTransactionDataInput]
     public var outputs: [ErgoTransactionOutput]
     public var size: Int?
-    public var numConfirmations: Int?
-    public init(_id: String? = nil, inputs: [ErgoTransactionInput]=[], dataInputs: [ErgoTransactionDataInput]=[], outputs: [ErgoTransactionOutput]=[], size: Int? = nil, numConfirmations: Int? = nil) {
+    public var numConfirmations: Int64?
+    public init(_id: String? = nil, inputs: [ErgoTransactionInput]=[], dataInputs: [ErgoTransactionDataInput]=[], outputs: [ErgoTransactionOutput]=[], size: Int? = nil, numConfirmations: Int64? = nil) {
         self.id = _id
         self.inputs = inputs
         self.dataInputs = dataInputs
